requirements.txt
# Recipe Editor - Python Dependencies

# Flask Web Framework
Flask==3.0.0

# WSGI HTTP Server for production (optional)
Werkzeug==3.0.1

# Template engine (included with Flask)
Jinja2==3.1.2

# URL routing (included with Flask)
itsdangerous==2.1.2

# Secure cookies (included with Flask)
click==8.1.7

# HTML/XML parsing (if needed in future)
MarkupSafe==2.1.3

# For development/testing (optional)
# python-dotenv==1.0.0  # Uncomment to use .env files for configuration

# Google Gemini AI for recipe extraction
google-genai==1.10.0

# For fetching web pages
requests==2.31.0
beautifulsoup4==4.12.3

# For PDF parsing
PyPDF2==3.0.1

# MySQL Database
mysql-connector-python==8.2.0
SQLAlchemy==2.0.23

# Password hashing
bcrypt==4.1.1

# Flask extensions for auth and sessions
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1

